Reviews added
Client is capable of adding reviews to a desired book. Client can also bookmark it and delete it and quality of life changes added such as better nav buttons

+++ app/src/main/java/com/example/bookcraftapplication/data/Review.kt
@@ -5,7 +5,7 @@ import com.example.bookcraft.data.Book
 
 data class Review(
     val bookName: String,
-    val rating: Int,
+    val rating: Float,
     val title: String,
     val description: String
 ) {
@@ -14,12 +14,12 @@ data class Review(
 
 
 var reviews = listOf(
-    Review("Something", 3, "Crap", "Its like the description says"),
-    Review("Something", 3, "Crap", "Its like the description says"),
-    Review("Something", 3, "Crap", "Its like the description says"),
-    Review("Something", 3, "Crap", "Its like the description says"),
-    Review("Something", 3, "Crap", "Its like the description says"),
-    Review("Something", 3, "Crap", "Its like the description says"),
-    Review("Something", 3, "Crap", "Its like the description says"),
-    Review("Something", 3, "Crap", "Its like the description says")
+    Review("Something", 3f, "Crap", "Its like the description says"),
+    Review("Something", 3f, "Crap", "Its like the description says"),
+    Review("Something", 3f, "Crap", "Its like the description says"),
+    Review("Something", 3f, "Crap", "Its like the description says"),
+    Review("Something", 3f, "Crap", "Its like the description says"),
+    Review("Something", 3f, "Crap", "Its like the description says"),
+    Review("Something", 3f, "Crap", "Its like the description says"),
+    Review("Something", 3f, "Crap", "Its like the description says")
 )
